+++ doc/CHANGES.md
@@ -2,6 +2,8 @@
 
 ## Highlighted Features
 
+* Prevent deletion of initial outgoing messages through the admin interface
+  (Gareth Rees)
 * Make it easier to find the "Resend" message button in the admin interface
   (Gareth Rees)
 * Install supported version of bundler through Rubygems on Debian Wheezy (Gareth